Dropdown for selecting which results to popup
Human only. Working on ADM
